Minor Refactoring from comments
